+++ lua/base/language.lua
@@ -129,7 +129,7 @@ end
 
 local function ClientLanguageOverride(client)
 	if client ~= nil then
-		local clientID = tostring(client:GetUserId())
+		local clientID = tonumber(client:GetUserId())
 		if clientID ~= nil then
 			if DAK.settings.clientlanguages[clientID] == nil or not GetIsLanguageValid(DAK.settings.clientlanguages[clientID]) then
 				DAK.settings.clientlanguages[clientID] = DAK.config.language.DefaultLanguage
@@ -206,7 +206,7 @@ end
 
 local function UpdateClientLanguageSetting(clientID, language)
 	if clientID ~= nil then
-		DAK.settings.clientlanguages[tostring(clientID)] = language
+		DAK.settings.clientlanguages[tonumber(clientID)] = language
 	end
 	DAK:SaveSettings()
 end